42 - remove icons and add title in homepage
