+++ public/manifest.webmanifest
@@ -1,15 +1,10 @@
 {
-  "name": "Nos Gestes Climat",
-  "description": "Calcule ton empreinte sur le climat",
+  "name": "Nos Gestes empreinte",
+  "description": "Enquête 2024",
   "display": "standalone",
   "lang": "fr",
   "orientation": "portrait",
   "theme_color": "#4949ba",
   "icons": [
-    {
-      "src": "/images/misc/logo.svg",
-      "sizes": "144x144",
-      "type": "image/svg+xml"
-    }
   ]
 }
